services modified to work with java

+++ models/message.go
@@ -6,8 +6,8 @@ import (
 
 type Message struct {
 	ID            uint64    `gorm:"primary_key" json:"id"`
-	FromId        int       `json:"from_id"`
-	ToId          int       `json:"to_id"`
+	FromId        uint64    `json:"from_id"`
+	ToId          uint64    `json:"to_id"`
 	MessageText   string    `json:"message_text"`
 	MessageStatus int       `json:"message_status"`
 	CreatedAt     time.Time `json:"created_at"`
